+++ src/components/button/Button.tsx
@@ -2,7 +2,6 @@ import React, {useState} from "react";
 import { useNavigate } from "react-router-dom";
 import '../../assets/css/button.css';
 
-
 interface ButtonProps {
     text: string;
     css: string;
@@ -17,8 +16,6 @@ function Button({text, css}: ButtonProps) {
             return "모니터링";
         else if (text === "back")
             return "이전으로";
-        else if (text === "start")
-            return "시작하기";
     };
     const goTo = () => {
         if (text === "home")
@@ -27,11 +24,9 @@ function Button({text, css}: ButtonProps) {
             navigate("/monitor");
         else if (text === "back")
             navigate(-1);
-        else if (text === "start")
-            navigate("/rendering");
     };
     return (
-            <button className={css} onClick={goTo}>{set_context()}</button>
+        <button className={css} type="submit" onClick={goTo}>{set_context()}</button>
     );
 }
